+++ src/Templates/default/footer.tpl.php
@@ -1,5 +1,5 @@
-	<?php foreach($javascript as $script): ?>
-	<script type="text/javascript" src="<?php print $script; ?>"></script>
+	<?php foreach ($javascript as $script): ?>
+		<script type="text/javascript" nonce="<?php print $nonce; ?>" src="<?php print $script; ?>"></script>
 	<?php endforeach; ?>
 	</body>
 </html>